Revert "feat: 레슨 관련 페이징 구조 변환"
This reverts commit b80805c74f9916a803702adfc57c54280e3f3bf8.

+++ src/main/java/com/threestar/trainus/domain/lesson/teacher/mapper/CreatedLessonMapper.java
@@ -27,7 +27,7 @@ public class CreatedLessonMapper {
 
 	// 개설한 레슨 목록과 총 레슨의 수를 응답 DTO로 변환
 	public static CreatedLessonListResponseDto toCreatedLessonListResponseDto(
-		List<Lesson> lessons, int totalCount) {
+		List<Lesson> lessons, Long totalCount) {
 
 		// 각 레슨을 DTO로 변환
 		List<CreatedLessonDto> lessonDtos = lessons.stream()
@@ -37,7 +37,7 @@ public class CreatedLessonMapper {
 		// 응답 DTO 생성
 		return CreatedLessonListResponseDto.builder()
 			.lessons(lessonDtos)
-			.count(totalCount)
+			.count(totalCount.intValue())
 			.build();
 	}
